描述：【后台】【教师端】课程学分数据类型转为double 日期：2020-02-03

+++ ulearning-spi/src/main/java/com/ky/ulearning/spi/system/dto/CourseDto.java
@@ -31,5 +31,5 @@ public class CourseDto extends BaseDto {
      * 学分
      */
     @ApiModelProperty("学分")
-    private Short courseCredit;
+    private Double courseCredit;
 }
